Extract MessageLogProcessor module
