+++ lib/line-builders.js
@@ -30,14 +30,6 @@ function appendSemicolon (lineParts) {
   }
 }
 
-function appendMessage (lineParts, { prettified }) {
-  const { prettifiedMessage } = prettified
-  if (prettifiedMessage) {
-    lineParts.push(' ')
-    lineParts.push(prettifiedMessage)
-  }
-}
-
 function appendEOL (lineParts, { EOL }) {
   if (lineParts.length > 0) {
     lineParts.push(EOL)
@@ -73,7 +65,6 @@ const builtInLineBuilders = {
   appendTime,
   appendMetadata,
   appendSemicolon,
-  appendMessage,
   appendEOL,
   appendObjectOrError
 }
