+++ lib/log-processors.js
@@ -1,4 +1,4 @@
-const { prettifyMessage, prettifyMetadata, prettifyTime } = require('./utils')
+const { prettifyMetadata, prettifyTime } = require('./utils')
 const { builtInLineBuilders } = require('./line-builders')
 const LogProcessor = require('./LogProcessor')
 const JsonLogProcessor = require('./processors/JsonLogProcessor')
@@ -6,9 +6,9 @@ const PrimitivesLogProcessor = require('./processors/PrimitivesLogProcessor')
 const SearchLogProcessor = require('./processors/SearchLogProcessor')
 const IgnoreLogProcessor = require('./processors/IgnoreLogProcessor')
 const LevelLogProcessor = require('./processors/LevelLogProcessor')
+const MessageLogProcessor = require('./processors/MessageLogProcessor')
 
 const {
-  appendMessage,
   appendMetadata,
   appendSemicolon,
   appendTime,
@@ -28,16 +28,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class MessageLogProcessor extends LogProcessor {
-  parse (input, context) {
-    return prettifyMessage(input, context)
-  }
-
-  build (lineParts, context) {
-    return appendMessage(lineParts, context)
   }
 }
