+++ lib/utils.js
@@ -2,7 +2,6 @@
 
 const dateformat = require('dateformat')
 const stringifySafe = require('fast-safe-stringify')
-const defaultColorizer = require('./colors')()
 const {
   DATE_FORMAT,
   ERROR_LIKE_KEYS,
@@ -14,7 +13,6 @@ const {
 module.exports = {
   isObject,
   prettifyErrorLog,
-  prettifyMessage,
   prettifyMetadata,
   prettifyObject,
   prettifyTime
@@ -158,29 +156,6 @@ function prettifyErrorLog ({
   }
 
   return result
-}
-
-/**
- * Prettifies a message string if the given `input` object has a message property.
- *
- * @param {object} input The log object which should have a message property.
- * @param {object} context The `context` object with the message to colorize.
- * @param {string} [context.messageKey='msg'] The property of the `input` that is the
- * message to be prettified.
- * @param {function} [context.colorizer] A colorizer function that has a
- * `.message(str)` method attached to it. This function should return a colorized
- * string which will be the "prettified" message. Default: a no-op colorizer.
- *
- * @returns {undefined|string} If the message key is not found, or the message
- * key is not a string, then `undefined` will be returned. Otherwise, a string
- * that is the prettified message.
- */
-function prettifyMessage (input, { prettified, messageKey = MESSAGE_KEY, colorizer = defaultColorizer } = {}) {
-  if (messageKey in input === false) return undefined
-  if (typeof input[messageKey] !== 'string') return undefined
-  const colorized = colorizer.message(input[messageKey])
-  prettified.prettifiedMessage = colorized
-  return undefined
 }
 
 /**
